+++ changelog.md
@@ -1,1 +1,1 @@
-- Added zooming
+- Fixed bug that players in adventure mode can break the image frame