+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.14.3
+- Fixed crash on servers